+++ src/main/kotlin/com/atlassian/performance/tools/virtualusers/api/browsers/Browser.kt
@@ -1,0 +1,5 @@
+package com.atlassian.performance.tools.virtualusers.api.browsers
+
+interface Browser {
+    fun start(): CloseableRemoteWebDriver
+}