+++ src/main/kotlin/com/atlassian/performance/tools/virtualusers/api/browsers/CloseableRemoteWebDriver.kt
@@ -1,0 +1,17 @@
+package com.atlassian.performance.tools.virtualusers.api.browsers
+
+import org.openqa.selenium.remote.RemoteWebDriver
+import java.io.Closeable
+
+open class CloseableRemoteWebDriver(
+    private val driver: RemoteWebDriver
+) : Closeable {
+
+    fun getDriver(): RemoteWebDriver {
+        return this.driver
+    }
+
+    override fun close() {
+        driver.quit()
+    }
+}